// client/components/Editor/types/enums.ts
export enum ElementTypeEnum {
  paragraph = 'paragraph',
  blockQuote = 'block-quote',
  listItem = 'list-item',
}

export enum MarkTypeEnum {
  bold = 'bold',
  italic = 'italic',
  underline = 'underline',
}

export enum ListTypeEnum {
  numberedList = 'numberedList',
  bulletedList = 'bulletedList',
}

export enum TextAlignTypes {
  left = 'left',
  right = 'right',
  center = 'center',
  justify = 'justify',
}

export enum ButtonIcons {
  formatBold = 'format_bold',
  formatItalic = 'format_italic',
  formatUnderlined = 'format_underlined',
  blockQuote = 'format_quote',
  numberedList = 'format_list_numbered',
  bulletedList = 'format_list_bulleted',
  left = 'format_align_left',
  center = 'format_align_center',
  right = 'format_align_right',
  justify = 'format_align_justify',
}